@@ -5,7 +5,7 @@ homepage: https://github.com/dart-bitcoin/bip32-dart
 author: anicdh <anic.dh@gmail.com>
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   pointycastle: ^3.0.1
@@ -13,4 +13,5 @@ dependencies:
   bs58check: ^1.0.2
 
 dev_dependencies:
-  test: ^1.0.0
+  flutter_test:
+    sdk: flutter
