@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import '../lib/bip32.dart' as bip32;
 import 'package:hex/hex.dart';
 
@@ -34,7 +36,7 @@ main() {
   print(childNeutered.toBase58());
   // => xpub6AvUGrnEpfvJ8L7GLRkBTByQ9uBvUHp9o5VxHrFxhvzV4dSWkySpNaBoLR9FpbnwRmTa69yLHF3QfcaxbWT7gWdwws5k4dpmJvqpEuMWwnj
 
-  bip32.BIP32 nodeFromSeed = bip32.BIP32.fromSeed(HEX.decode("000102030405060708090a0b0c0d0e0f"));
+  bip32.BIP32 nodeFromSeed = bip32.BIP32.fromSeed(Uint8List.fromList(HEX.decode("000102030405060708090a0b0c0d0e0f")));
   print(nodeFromSeed.toBase58());
   // => xprv9s21ZrQH143K3QTDL4LXw2F7HEK3wJUD2nW2nRk4stbPy6cq3jPPqjiChkVvvNKmPGJxWUtg6LnF5kejMRNNU3TGtRBeJgk33yuGBxrMPHi
 
@@ -42,7 +44,7 @@ main() {
   print(nodeFromPub.toBase58());
   // => xpub661MyMwAqRbcFtXgS5sYJABqqG9YLmC4Q1Rdap9gSE8NqtwybGhePY2gZ29ESFjqJoCu1Rupje8YtGqsefD265TMg7usUDFdp6W1EGMcet8
 
-  var message = HEX.decode("0202020202020202020202020202020202020202020202020202020202020202");
+  var message = Uint8List.fromList(HEX.decode("0202020202020202020202020202020202020202020202020202020202020202"));
   var signature = nodeFromSeed.sign(message);
   print(signature);
   // => [63, 219, 20, 114, 95, 184, 192, 55, 216, 206, 126, 121, 17, 71, 64, 70, 163, 82, 247, 73, 243, 95, 30, 137, 177, 155, 100, 225, 177, 203, 217, 147, 122, 64, 208, 129, 54, 133, 113, 41, 216, 160, 191, 15, 136, 98, 235, 25, 219, 178, 70, 222, 127, 151, 135, 242, 25, 192, 161, 187, 187, 84, 81, 215]
